+++ app/src/pages/Insights/Insight.js
@@ -21,8 +21,8 @@ import mediumDraftImporter from 'medium-draft/lib/importer'
 import 'medium-draft/lib/index.css'
 import InsightsLayout from './InsightsLayout'
 import Panel from './../../components/Panel'
-import LikeBtn from './../EventVotesNew/LikeBtn'
-import { votePostGQL, unvotePostGQL } from './../EventVotes'
+import LikeBtn from './../InsightsNew/LikeBtn'
+import { votePostGQL, unvotePostGQL } from './../InsightsPage'
 import { getBalance } from './../UserSelectors'
 import './Insight.css'
