+++ app/src/pages/InsightsNew/ConfirmNewInsight.js
