+++ app/src/pages/InsightsNew/CreateInsight.css
