+++ app/src/pages/InsightsNew/InsightsNew.css
