+++ app/src/pages/InsightsPage.css
